# api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Yungdrung",
  "description": "Трекер задач с последовательными шагами и контрольным временем. Пишет только ядро, считает только ядро; оболочка показывает то, что пришло (CONTRACT.md).",
  "version": "1"
 },
 "paths": {
  "/api/v1/feed": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "Feed",
    "operationId": "feed_api_v1_feed_get",
    "parameters": [
     {
      "name": "now",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Now"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FeedResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/backlog": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "Backlog",
    "operationId": "backlog_api_v1_backlog_get",
    "parameters": [
     {
      "name": "now",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Now"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BacklogResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tasks/{task_id}/steps/{step_id}/mark": {
   "post": {
    "tags": [
     "v1"
    ],
    "summary": "Mark",
    "operationId": "mark_api_v1_tasks__task_id__steps__step_id__mark_post",
    "parameters": [
     {
      "name": "task_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Task Id"
      }
     },
     {
      "name": "step_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Step Id"
      }
     },
     {
      "name": "now",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Now"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MarkIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MarkResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tasks/{task_id}/steps/{step_id}/undo": {
   "post": {
    "tags": [
     "v1"
    ],
    "summary": "Undo",
    "operationId": "undo_api_v1_tasks__task_id__steps__step_id__undo_post",
    "parameters": [
     {
      "name": "task_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Task Id"
      }
     },
     {
      "name": "step_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Step Id"
      }
     },
     {
      "name": "now",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Now"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MarkResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/parse-date": {
   "post": {
    "tags": [
     "v1"
    ],
    "summary": "Parse Date",
    "operationId": "parse_date_api_v1_parse_date_post",
    "parameters": [
     {
      "name": "now",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Now"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TextIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WhenResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/extract-when": {
   "post": {
    "tags": [
     "v1"
    ],
    "summary": "Extract When",
    "operationId": "extract_when_api_v1_extract_when_post",
    "parameters": [
     {
      "name": "now",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Now"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TextIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExtractResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/search": {
   "get": {
    "tags": [
     "v1"
    ],
    "summary": "Search",
    "operationId": "search_api_v1_search_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "BacklogResult": {
    "properties": {
     "now": {
      "type": "string",
      "title": "Now"
     },
     "backlog": {
      "items": {
       "$ref": "#/components/schemas/FeedRow"
      },
      "type": "array",
      "title": "Backlog"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     },
     "broken": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Broken"
     }
    },
    "type": "object",
    "required": [
     "now",
     "backlog",
     "count"
    ],
    "title": "BacklogResult"
   },
   "Counts": {
    "properties": {
     "overdue": {
      "type": "integer",
      "title": "Overdue"
     },
     "today": {
      "type": "integer",
      "title": "Today"
     },
     "waiting": {
      "type": "integer",
      "title": "Waiting"
     }
    },
    "type": "object",
    "required": [
     "overdue",
     "today",
     "waiting"
    ],
    "title": "Counts"
   },
   "ExtractResult": {
    "properties": {
     "title": {
      "type": "string",
      "title": "Title"
     },
     "date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "past": {
      "type": "boolean",
      "title": "Past",
      "default": false
     },
     "span": {
      "anyOf": [
       {
        "prefixItems": [
         {
          "type": "integer"
         },
         {
          "type": "integer"
         }
        ],
        "type": "array",
        "maxItems": 2,
        "minItems": 2
       },
       {
        "type": "null"
       }
      ],
      "title": "Span"
     }
    },
    "type": "object",
    "required": [
     "title"
    ],
    "title": "ExtractResult",
    "description": "Быстрый ввод одной строкой: название, дата и границы распознанного куска\nв исходном тексте — для подсветки и вырезания из названия."
   },
   "FeedResult": {
    "properties": {
     "now": {
      "type": "string",
      "title": "Now"
     },
     "feed": {
      "items": {
       "$ref": "#/components/schemas/FeedRow"
      },
      "type": "array",
      "title": "Feed"
     },
     "overdue_count": {
      "type": "integer",
      "title": "Overdue Count"
     },
     "counts": {
      "$ref": "#/components/schemas/Counts"
     },
     "next_ahead": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FeedRow"
       },
       {
        "type": "null"
       }
      ]
     },
     "stalled_count": {
      "type": "integer",
      "title": "Stalled Count"
     },
     "broken": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Broken"
     }
    },
    "type": "object",
    "required": [
     "now",
     "feed",
     "overdue_count",
     "counts",
     "next_ahead",
     "stalled_count"
    ],
    "title": "FeedResult",
    "description": "Лента «Что сегодня». Просроченное в строки не попадает: по 6.1 оно живёт\nотдельной плашкой — пятнадцать красных строк парализуют экран."
   },
   "FeedRow": {
    "properties": {
     "task_id": {
      "type": "integer",
      "title": "Task Id"
     },
     "task": {
      "type": "string",
      "title": "Task"
     },
     "step": {
      "type": "integer",
      "title": "Step"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "group": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Group"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "control_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Control At"
     },
     "show_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Show At"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "postponed": {
      "type": "integer",
      "title": "Postponed"
     },
     "stalled": {
      "type": "boolean",
      "title": "Stalled"
     },
     "tags": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Tags"
     },
     "last_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Reason"
     },
     "actions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Actions"
     }
    },
    "type": "object",
    "required": [
     "task_id",
     "task",
     "step",
     "title",
     "control_at",
     "show_at",
     "state",
     "postponed",
     "stalled",
     "tags",
     "last_reason",
     "actions"
    ],
    "title": "FeedRow",
    "description": "Строка ленты или завала. Всё вычислено здесь: оболочка только показывает.\nРаздел 6.1 ТЗ перечисляет, что видно в строке: название шага, название\nзадачи, время контроля, теги, счётчик переносов, если больше нуля."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "MarkIn": {
    "properties": {
     "op": {
      "type": "string",
      "title": "Op"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "To"
     }
    },
    "type": "object",
    "required": [
     "op"
    ],
    "title": "MarkIn",
    "description": "Отметка шага. `to` — человеческий ввод («завтра в 9», «+3», ISO):\nразбирает ядро, оболочка дату не трогает (`CONTRACT.md`)."
   },
   "MarkResult": {
    "properties": {
     "ok": {
      "type": "boolean",
      "title": "Ok",
      "default": true
     },
     "op": {
      "type": "string",
      "title": "Op"
     },
     "task_id": {
      "type": "integer",
      "title": "Task Id"
     },
     "task": {
      "type": "string",
      "title": "Task"
     },
     "step": {
      "type": "integer",
      "title": "Step"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "task_status": {
      "type": "string",
      "title": "Task Status"
     },
     "next_step_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Step Id"
     },
     "next_step_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Step Title"
     },
     "dates_assigned": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Dates Assigned"
     },
     "next_check": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Check"
     },
     "stalled": {
      "type": "integer",
      "title": "Stalled",
      "default": 0
     },
     "hint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hint"
     },
     "undone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Undone"
     },
     "control_date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Control Date"
     },
     "row": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FeedRow"
       },
       {
        "type": "null"
       }
      ]
     },
     "counts": {
      "$ref": "#/components/schemas/Counts"
     }
    },
    "type": "object",
    "required": [
     "op",
     "task_id",
     "task",
     "step",
     "status",
     "task_status",
     "counts"
    ],
    "title": "MarkResult",
    "description": "Ответ на отметку шага. Несёт новое состояние — обновлённую строку ленты\n(`row`, None если шаг из ленты и завала ушёл) и счётчики, — чтобы клиенту\nне нужен был второй круг до сервера. Это контрактно чистая версия\nоптимистичного интерфейса: оболочка по-прежнему ничего не вычисляет."
   },
   "SearchResult": {
    "properties": {
     "ok": {
      "type": "boolean",
      "title": "Ok",
      "default": true
     },
     "query": {
      "type": "string",
      "title": "Query"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     },
     "results": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Results"
     }
    },
    "type": "object",
    "required": [
     "query",
     "count",
     "results"
    ],
    "title": "SearchResult"
   },
   "TextIn": {
    "properties": {
     "text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Text"
     }
    },
    "type": "object",
    "title": "TextIn"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WhenResult": {
    "properties": {
     "ok": {
      "type": "boolean",
      "title": "Ok"
     },
     "date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "past": {
      "type": "boolean",
      "title": "Past",
      "default": false
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "type": "object",
    "required": [
     "ok"
    ],
    "title": "WhenResult",
    "description": "Во что превратился человеческий ввод даты — подсказка под полем формы.\nПодпись («завтра в 09:30», «18.10.2026, через 40 дн.») считается здесь,\nа не в браузере: по контракту оболочка не пересказывает данные ядра\nсвоими словами. Нераспознанный ввод — это ответ, а не ошибка запроса:\nполе опрашивается на каждое нажатие клавиши."
   }
  }
 }
}